Add RTK availability, improve file structure
