+++ message_converter/package.xml
@@ -54,16 +54,19 @@
   <build_depend>libsbp_ros_msgs</build_depend>
   <build_depend>piksi_rtk_msgs</build_depend>
   <build_depend>roscpp</build_depend>
+  <build_depend>std_msgs</build_depend>
   <build_export_depend>geometry_msgs</build_export_depend>
   <build_export_depend>gnss_msgs</build_export_depend>
   <build_export_depend>libsbp_ros_msgs</build_export_depend>
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>piksi_rtk_msgs</build_export_depend>
+  <build_export_depend>std_msgs</build_export_depend>
   <exec_depend>geometry_msgs</exec_depend>
   <exec_depend>gnss_msgs</exec_depend>
   <exec_depend>libsbp_ros_msgs</exec_depend>
   <exec_depend>roscpp</exec_depend>
   <exec_depend>piksi_rtk_msgs</exec_depend>
+  <exec_depend>std_msgs</exec_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
